All variable type trees are now cached (before values are read) Pointers no longer directly read their pointee types, but read them from cache Big code refactor for Variables
